+++ lib/yt/collections/channels.rb
@@ -9,6 +9,12 @@ module Yt
 
     private
 
+      def attributes_for_new_item(data)
+        super(data).tap do |attributes|
+          attributes[:statistics] = data['statistics']
+        end
+      end
+
       # @return [Hash] the parameters to submit to YouTube to list channels.
       # @see https://developers.google.com/youtube/v3/docs/channels/list
       def list_params
@@ -16,7 +22,9 @@ module Yt
       end
 
       def channels_params
-        resources_params.merge mine: true
+        params = resources_params
+        params.merge! mine: true if @parent
+        apply_where_params! params
       end
     end
   end