+++ lib/yt/collections/partnered_channels.rb
@@ -4,7 +4,34 @@ module Yt
   module Collections
     class PartneredChannels < Channels
 
+      def includes(relationship)
+        self.tap do
+          @items = []
+          @includes_relationship = relationship
+        end
+      end
+
     private
+
+      def attributes_for_new_item(data)
+        super(data).tap do |attributes|
+          attributes[:viewer_percentages] = data['viewerPercentages']
+        end
+      end
+
+      def eager_load_items_from(items)
+        if @includes_relationship == :viewer_percentages
+          filters = "channel==#{items.map{|item| item['id']}.join(',')}"
+          ids = "contentOwner==#{@auth.owner_name}"
+          conditions = {ids: ids, filters: filters}
+          viewer_percentages = Collections::ViewerPercentages.new auth: @auth
+          viewer_percentages = viewer_percentages.where conditions
+          items.each do |item|
+            item['viewerPercentages'] = viewer_percentages[item['id']]
+          end
+        end
+        items
+      end
 
       # @private
       # @note Partnered Channels overwrites +channel_params+ since the query
@@ -13,7 +40,7 @@ module Yt
         super.tap do |params|
           params.delete :mine
           params[:managed_by_me] = true
-          params[:on_behalf_of_content_owner] = @parent.owner_name
+          params[:on_behalf_of_content_owner] = @auth.owner_name
         end
       end