+++ CHANGELOG.md
@@ -6,6 +6,12 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
+## 0.13.6 - 2014-10-08
+
+* [ENHANCEMENT] Accept `includes(:viewer_percentages)` in `.partnered_channels` to eager-load multiple viewer percentages at once.
+* [ENHANCEMENT] Accept `where` in ViewerPercentages to collect data for multiple channels at once.
+* [ENHANCEMENT] Accept `part` in the `where` clause of Channels, so statistics can be loaded at once.
+
 ## 0.13.5 - 2014-10-06
 
 * [ENHANCEMENT] Add `advertising_options_set` and `ad_formats` to video